Continue work on logger and fixing Clang Static Analyzer's warnings

+++ Engine/scenes/level/lvl_player.h
@@ -271,7 +271,7 @@ class LVL_Player :
         /*---------------------*/
 
         virtual void lua_onInit() {}
-        virtual void lua_onLoop(float) {}
+        virtual void lua_onLoop(double) {}
         virtual void lua_onHarm(LVL_Player_harm_event *) {}
         virtual void lua_onTransform(unsigned long, unsigned long) {}
         virtual void lua_onTakeNpc(LVL_Npc *) {}
